Updated DynamoDB portion to access 'standalone' DynamoDB. This includes both the connection configuration and the DynamoDB Object Mapper
Created an activity to allow the user to upload photos. The activity prompts the user to select a photo from their library, then displays that photo and allows it to be uploaded to S3 storage.

Cleaned out a lot of unnecessary files and trimmed portions of used files

*** IMPORTANT ***
S3 Upload will not work without access KEY and SECRET